Allow disabling the Commercial Support message
When including Atmosphere as an implementation detail in a product that
provides its own commercial support, the message printed during Atmosphere
initialization might make the user believe that async-io.org provides
commercial support for the entire product, which is obviously not the case.

This change lets integrators disable the message to avoid confusion.