adapts AtmosphereConfigurator to avoid RuntimeExceptions in OSGi
When using e.g. Jetty 10 in an Eclipse equinox OSGi environment the
java.util.ServiceLoader does not work. This results in a
RuntimeException during initialization of the javax.websocket
implementation when calling the methods checkOrigin,
getNegotiatedSubprotocol or getNegotiatedExtensions. The change checks
if the ServiceLoader finds a default implementation of
ServerEndpointConfig.Configurator and calls the super class methods if
so. Otherwise a simple default implementation for all 3 methods is
provided.